+++ .github/workflows/ci.yml
@@ -13,12 +13,15 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Install Dependencies
+        working-directory: ./ConcordiaMaps
         run: npm install
 
       - name: Lint Code
+        working-directory: ./ConcordiaMaps
         run: npm run lint
 
       - name: Format Code
+        working-directory: ./ConcordiaMaps
         run: npx prettier --check .
 
   test:
@@ -28,12 +31,15 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Install Dependencies
+        working-directory: ./ConcordiaMaps
         run: npm install
 
       - name: Lint Code
+        working-directory: ./ConcordiaMaps
         run: npm run lint
 
       - name: Format Code
+        working-directory: ./ConcordiaMaps
         run: npx prettier --check .
 
   deploy:
