+++ ConcordiaMaps/App.js
@@ -1,10 +1,8 @@
 import * as React from "react";
-import { StyleSheet } from "react-native";
-// import { Text} from 'react-native';
-
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import HomeScreen from "./screen/HomeScreen";
+import styles from "./styles";
 
 const Stack = createNativeStackNavigator();
 
@@ -21,11 +19,3 @@ export default function App() {
     </NavigationContainer>
   );
 }
-const styles = StyleSheet.create({
-  container: {
-    width: "100%",
-    height: "100%",
-    backgroundColor: "#912338",
-    color: "blue",
-  },
-});
