+++ ConcordiaMaps/components/Header.js
@@ -1,5 +1,6 @@
 import React from "react";
-import { SafeAreaView, StyleSheet, View, Text, Image } from "react-native";
+import { SafeAreaView, View, Text, Image } from "react-native";
+import styles from "../styles";
 
 function Header() {
   return (
@@ -14,30 +15,5 @@ function Header() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  header: {
-    width: "100%",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#912338",
-    padding: 10,
-    position: "relative",
-    zIndex: 1, // Ensures the logo appears on top
-  },
-  logo: {
-    width: 50,
-    height: 50,
-    marginLeft: 20,
-  },
-  headerText: {
-    fontWeight: "bold",
-    fontFamily: "Times New Roman",
-    fontSize: 25,
-    color: "white",
-    marginLeft: 10,
-  },
-});
 
 export default Header;
