+++ ConcordiaMaps/screen/HomeScreen.js
@@ -1,18 +1,22 @@
-import React, { useState, useEffect, useRef } from "react";
-import { View, Text, StyleSheet, TouchableOpacity, Image } from "react-native";
+import React from "react";
+import {useState, useEffect, useRef} from "react";
+import { View, StyleSheet, Text, TouchableOpacity, Image } from "react-native";
 import axios from "axios";
 import MapView from "react-native-maps";
 import NavBar from "../components/NavBar";
 import Header from "../components/Header";
+import styles from "../styles";
 
 function HomeScreen() {
-  const [postalCode, setPostalCode] = useState('H3G 1M8');
-  const [coordinates, setCoordinates] = useState(null);
-  const [error, setError] = useState('');
-  const mapRef = useRef(null);
   const loyolaPostalCode = 'H4B 1R6';
   const sgwPostalCode = 'H3G 1M8';
 
+  const [postalCode, setPostalCode] = useState(sgwPostalCode);
+  const [coordinates, setCoordinates] = useState(null);
+  const [error, setError] = useState('');
+
+  const mapRef = useRef(null);
+  
   const convertToCoordinates = async (postal_code) => {
     const key = "AIzaSyAW8gOP1PJiZp1br3kOPSlRYdPlDoGkkR4"; // Replace with your actual API key
     try {
@@ -38,6 +42,7 @@ function HomeScreen() {
       setError('Something went wrong. Please try again later.');
     }
   };
+
   useEffect(() => {
     if (coordinates && mapRef.current) {
       mapRef.current.animateToRegion({
@@ -45,7 +50,7 @@ function HomeScreen() {
         longitude: coordinates.longitude,
         latitudeDelta: 0.01,
         longitudeDelta: 0.01,
-      }, 2900); // Duration of the animation in milliseconds
+      }, 2500); // Duration of the animation in milliseconds
     }
   }, [coordinates]);
   
@@ -60,54 +65,38 @@ function HomeScreen() {
   return (
     <View style={styles.container}>
       <Header />
-      <View style={styles.mapContainer}>
-        {coordinates ? (
+      <NavBar /> 
+          {coordinates ? (
           <>
-            <MapView
-              ref = {mapRef}
-              style={styles.map}
-              initialRegion={{
-                latitude: coordinates.latitude,
-                longitude: coordinates.longitude,
-                latitudeDelta: 0.01, // Adjusted for a balance between close zoom and larger area
-                longitudeDelta: 0.01,
-              }}
-            />
-          </>
-        ) : (
-          <Text>Loading...</Text>
-        )}
-        {error ? <Text>Error: {error}</Text> : null}
-        <TouchableOpacity
+            <MapView 
+            style={styles.map} 
+            ref={mapRef}
+            initialRegion={{
+              latitude: coordinates.latitude,
+              longitude: coordinates.longitude,
+              latitudeDelta: 0.01, // Adjusted for a balance between close zoom and larger area
+              longitudeDelta: 0.01,
+            }}/>
+          </>) 
+              :
+            (<Text>Loading...</Text>)
+          }    
+          {error ? <Text>Error: {error}</Text> : null}
+          <TouchableOpacity
           onPress={handleChangeCampuses}
           activeOpacity={0.7}
-          style={styles.button}
+          style={localStyles.button}
         >
           <Image
-            style={styles.buttonImage}
+            style={localStyles.buttonImage}
             source={require('../assets/download.jpg')}
             resizeMode={'cover'} // cover or contain its up to you view look
           />
         </TouchableOpacity>
-      </View>
-      <NavBar />
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  mapContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  map: {
-    width: '100%',
-    height: '100%',
-  },
+const localStyles = StyleSheet.create({
   button: {
     position: 'absolute',
     bottom: 20,
@@ -120,5 +109,4 @@ const styles = StyleSheet.create({
     height: '100%',
   },
 });
-
 export default HomeScreen;
