--- package.json
@@ -1,35 +1,0 @@
-{
-  "name": "concordiamaps",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "lint": "eslint . --ignore-pattern node_modules/"
-  },
-  "dependencies": {
-    "@react-native-async-storage/async-storage": "^2.1.0",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/native-stack": "^7.2.0",
-    "expo": "~52.0.27",
-    "expo-status-bar": "~2.0.1",
-    "react": "18.3.1",
-    "react-native": "0.76.6",
-    "react-native-gesture-handler": "^2.22.1",
-    "react-native-maps": "^1.20.1",
-    "react-native-reanimated": "^3.16.7",
-    "react-native-safe-area-context": "^4.12.0",
-    "react-native-screens": "~4.4.0",
-    "react-native-vector-icons": "^10.2.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@eslint/js": "^9.19.0",
-    "eslint": "^9.19.0",
-    "eslint-plugin-react": "^7.37.4",
-    "globals": "^15.14.0"
-  },
-  "private": true
-}
